Hacky fix to search issues
Hard-coded that player_id (instead of player) needs to be different in the search. Seems correct as well - previously, the state copies referenced different players.

Multi-ply search still does not work, probably due to the fact that moves are not grouped, so picking up pieces (which is a free move) will not add to the depth.